+++ evm/evm-typegen/src/main.ts
@@ -30,6 +30,10 @@ The generated facades are assumed to be used by "squids" indexing EVM data.
             'etherscan API to fetch contract ABI by a known address',
             validator.Url(['http:', 'https:'])
         )
+        .option(
+            '--etherscan-api-key <key>',
+            'etherscan API key'
+        )
         .option('--clean', 'delete output directory before run')
         .addHelpText('afterAll', `
 ABI file can be specified in three ways:
@@ -58,6 +62,7 @@ squid-evm-typegen src/abi 0xBB9bc244D798123fDe783fCc1C72d3Bb8C189413#contract
         clean?: boolean,
         multicall?: boolean,
         etherscanApi?: string
+        etherscanApiKey?: string
     }
     let dest = new OutDir(program.processedArgs[0])
     let specs = program.processedArgs[1] as Spec[]
@@ -89,9 +94,9 @@ squid-evm-typegen src/abi 0xBB9bc244D798123fDe783fCc1C72d3Bb8C189413#contract
 }, err => LOG.fatal(err))
 
 
-async function read(spec: Spec, options?: {etherscanApi?: string}): Promise<any> {
+async function read(spec: Spec, options?: {etherscanApi?: string, etherscanApiKey?: string}): Promise<any> {
     if (spec.kind == 'address') {
-        return fetchFromEtherscan(spec.src, options?.etherscanApi)
+        return fetchFromEtherscan(spec.src, options?.etherscanApi, options?.etherscanApiKey)
     }
     let abi: any
     if (spec.kind == 'url') {
@@ -109,17 +114,22 @@ async function read(spec: Spec, options?: {etherscanApi?: string}): Promise<any>
 }
 
 
-async function fetchFromEtherscan(address: string, api?: string): Promise<any> {
+async function fetchFromEtherscan(address: string, api?: string, apiKey?: string): Promise<any> {
     api = api || 'https://api.etherscan.io/'
     let url = new URL('api?module=contract&action=getabi', api)
     url.searchParams.set('address', address)
+    if (apiKey) {
+        url.searchParams.set('apiKey', apiKey)
+    }
     let response: {status: string, result: string}
-    let attempts = 2
+    let attempts = 0
     while (true) {
         response = await GET(url.toString())
-        if (response.status == '0' && response.result.includes('rate limit') && --attempts) {
-            LOG.warn('faced rate limit error while trying to fetch contract ABI. Trying again in 2 seconds.')
-            await wait(2000)
+        if (response.status == '0' && response.result.includes('rate limit') && attempts < 4) {
+            attempts += 1
+            let timeout = attempts * 2
+            LOG.warn(`faced rate limit error while trying to fetch contract ABI. Trying again in ${timeout} seconds.`)
+            await wait(timeout * 1000)
         } else {
             break
         }