+++ substrate/frontier/src/index.ts
@@ -1,2 +1,3 @@
 export {EvmLog, getEvmLog} from './log'
+export {TransactionResult, getTransactionResult} from './executed'
 export {Transaction, LegacyTransaction, EIP1559Transaction, EIP2930Transaction, getTransaction} from './transaction'