+++ substrate/frontier/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @subsquid/frontier
 
-This log was last generated on Thu, 09 Feb 2023 15:10:00 GMT and should not be manually modified.
+This log was last generated on Sun, 19 Feb 2023 09:15:04 GMT and should not be manually modified.
+
+## 0.2.0
+Sun, 19 Feb 2023 09:15:04 GMT
+
+### Minor changes
+
+- introduce `getTransactionResult()` to extract tx status from `Ethereum.Executed` event
 
 ## 0.1.2
 Thu, 09 Feb 2023 15:10:00 GMT